refactor: flytt TemplatesPageContent og RoleplayEncounterScreens til feature-mapper (T-13, T-14)
TemplatesPageContent.tsx, templatesPageUtils.ts og tilhorende test er
flyttet til apps/web/src/features/templates/. RoleplayEncounterScreens.tsx,
tilhorende test og alle seks komponenter er flyttet til
apps/web/src/features/roleplay/. Rene flytt -- ingen logikkendring.
Importstier i page.tsx, EncounterDetail.tsx og CampaignWorkspaceShell.tsx
er oppdatert til @/-aliaser. 305 tester gronne.

Closes #169, #170
