+++ apps/web/app/(app)/campaigns/[campaignId]/CampaignWorkspaceShell.tsx
@@ -30,7 +30,7 @@ import CharacterWorkspacePanel from "./CharacterWorkspacePanel";
 import ScenarioDetailPageContent from "./scenarios/[scenarioId]/ScenarioDetailPageContent";
 import ScenarioPlayerPageContent from "./scenarios/[scenarioId]/player/ScenarioPlayerPageContent";
 import ScenarioPlayerCombatPageContent from "./scenarios/[scenarioId]/player/combat/ScenarioPlayerCombatPageContent";
-import { PlayerRoleplayingEncounterScreen } from "../../encounters/[id]/RoleplayEncounterScreens";
+import { PlayerRoleplayingEncounterScreen } from "@/features/roleplay/RoleplayEncounterScreens";
 
 interface CampaignWorkspaceShellProps {
   campaignId: string;
