+++ apps/web/app/(app)/templates/page.tsx
@@ -1,4 +1,4 @@
-import TemplatesPageContent from "./TemplatesPageContent";
+import TemplatesPageContent from "@/features/templates/TemplatesPageContent";
 
 export default function TemplatesPage() {
   return <TemplatesPageContent />;
