+++ apps/web/src/features/roleplay/components/RoleplayCalculationPanel.tsx
