+++ apps/web/src/features/roleplay/components/RoleplayRollBlock.tsx
